+++ js/gameRecordSkeleton.js
@@ -1,7 +1,9 @@
-// gameRecord 會被 CreateNewGame() 指定指向新的物件
+// gameRecord 無法被其他module影響而指向新物件，它可以變更該記憶體裡的property
 
-export let gameRecord = {'id':null,'targetNum':null,'guessedTimes':0,'answerHistory':[]};
+export const gameRecord = {'id':null,'targetNum':null,'guessedTimes':0,'answerHistory':[]};
 
-export const gameRecordHistory = {'timesOfPlaying':0 ,'RecordsForEachGame':[]};
+export const recordForEachGame = [];
 
 export const gameStatus = {'status':'notReady'}
+
+export let  snapshotOfHistory = [];